requirements.txt
streamlit
langchain
langchain-anthropic
langchain-openai
langchain-groq
groq
langgraph
chromadb
python-chess
pydantic-settings
pydantic>=2.0
python-dotenv
openai
pandas
rank-bm25
